requirements.txt
requests
simple-salesforce
supabase
python-dotenv
modal
